feat: add Deno Deploy provider output support
