+++ packages/agent/src/stream-output.ts
@@ -8,10 +8,83 @@ interface FinalizedStreamOutput<T> {
   value: ReadableStream<T>
 }
 
+interface AgentUIMessageStreamWriter {
+  write(event: unknown): void
+}
+
+const uiMessageStreamHeaders = {
+  "cache-control": "no-cache",
+  connection: "keep-alive",
+  "content-type": "text/event-stream",
+  "x-accel-buffering": "no",
+  "x-vercel-ai-ui-message-stream": "v1",
+} as const
+
 export function isUIMessageStreamResult(value: unknown): value is { toUIMessageStream: () => ReadableStream<unknown> } {
   return typeof value === "object"
     && value !== null
     && typeof (value as { toUIMessageStream?: unknown }).toUIMessageStream === "function"
+}
+
+export function createAgentUIMessageStream(options: {
+  execute: (context: { writer: AgentUIMessageStreamWriter }) => MaybePromise<void>
+}): ReadableStream<unknown> {
+  return new ReadableStream<unknown>({
+    start(controller) {
+      const writer: AgentUIMessageStreamWriter = {
+        write(event) {
+          try {
+            controller.enqueue(event)
+          }
+          catch {
+            // The consumer may cancel the response before the agent finishes.
+          }
+        },
+      }
+      void Promise.resolve(options.execute({ writer }))
+        .catch((error) => {
+          try {
+            controller.enqueue({
+              errorText: error instanceof Error ? error.message : "Agent stream failed.",
+              type: "error",
+            })
+          }
+          catch {}
+        })
+        .finally(() => {
+          try {
+            controller.close()
+          }
+          catch {}
+        })
+    },
+  })
+}
+
+export function createAgentUIMessageStreamResponse(options: {
+  headers?: ConstructorParameters<typeof Headers>[0]
+  status?: number
+  statusText?: string
+  stream: ReadableStream<unknown>
+}): Response {
+  const headers = new Headers(options.headers)
+  for (const [key, value] of Object.entries(uiMessageStreamHeaders)) {
+    if (!headers.has(key)) headers.set(key, value)
+  }
+  const encoder = new TextEncoder()
+  const stream = options.stream.pipeThrough(new TransformStream<unknown, Uint8Array>({
+    flush(controller) {
+      controller.enqueue(encoder.encode("data: [DONE]\n\n"))
+    },
+    transform(part, controller) {
+      controller.enqueue(encoder.encode(`data: ${JSON.stringify(part)}\n\n`))
+    },
+  }))
+  return new Response(stream, {
+    headers,
+    status: options.status,
+    statusText: options.statusText,
+  })
 }
 
 export function withReadableStreamCleanup<T>(stream: ReadableStream<T>, cleanup: (error?: unknown) => Promise<void>): ReadableStream<T> {
@@ -66,40 +139,40 @@ function uiDataType(data: unknown): `data-${string}` {
   return `data-${type || "event"}`
 }
 
-async function writeEventsToUiMessageStream(writer: { write: (event: never) => void }, events: AsyncIterable<unknown>) {
+async function writeEventsToUiMessageStream(writer: AgentUIMessageStreamWriter, events: AsyncIterable<unknown>) {
   const messageId = crypto.randomUUID()
   let textStarted = false
   let finished = false
-  writer.write({ type: "start", messageId } as never)
+  writer.write({ type: "start", messageId })
   for await (const event of events) {
     const type = streamEventType(event)
     if (type === "text-delta") {
       const text = (event as { delta?: unknown, text?: unknown }).text ?? (event as { delta?: unknown }).delta
       if (typeof text !== "string" || !text) continue
       if (!textStarted) {
-        writer.write({ type: "text-start", id: messageId } as never)
+        writer.write({ type: "text-start", id: messageId })
         textStarted = true
       }
-      writer.write({ type: "text-delta", id: messageId, delta: text } as never)
+      writer.write({ type: "text-delta", id: messageId, delta: text })
       continue
     }
     if (type === "tool-call") {
       const tool = event as { id?: unknown, input?: unknown, name?: unknown }
-      writer.write({ type: "tool-input-available", toolCallId: tool.id, toolName: tool.name, input: tool.input } as never)
+      writer.write({ type: "tool-input-available", toolCallId: tool.id, toolName: tool.name, input: tool.input })
       continue
     }
     if (type === "tool-result") {
       const tool = event as { error?: unknown, id?: unknown, name?: unknown, output?: unknown }
       if (typeof tool.error === "string") {
-        writer.write({ type: "tool-output-error", toolCallId: tool.id, toolName: tool.name, errorText: tool.error } as never)
+        writer.write({ type: "tool-output-error", toolCallId: tool.id, toolName: tool.name, errorText: tool.error })
         continue
       }
-      writer.write({ type: "tool-output-available", toolCallId: tool.id, toolName: tool.name, output: tool.output } as never)
+      writer.write({ type: "tool-output-available", toolCallId: tool.id, toolName: tool.name, output: tool.output })
       continue
     }
     if (type === "data") {
       const data = (event as { data?: unknown }).data
-      writer.write({ type: uiDataType(data), data } as never)
+      writer.write({ type: uiDataType(data), data })
       continue
     }
     if (type === "finish") {
@@ -111,8 +184,8 @@ async function writeEventsToUiMessageStream(writer: { write: (event: never) => v
       throw error.error || new Error(typeof error.message === "string" ? error.message : "Agent stream failed.")
     }
   }
-  if (textStarted) writer.write({ type: "text-end", id: messageId } as never)
-  writer.write({ type: "finish", finishReason: finished ? "stop" : "unknown" } as never)
+  if (textStarted) writer.write({ type: "text-end", id: messageId })
+  writer.write({ type: "finish", finishReason: finished ? "stop" : "unknown" })
 }
 
 export async function finalizeUiMessageStreamOutput(
@@ -129,10 +202,10 @@ export async function finalizeUiMessageStreamOutput(
   const stream = hasUiMessageStream
     ? rendered.toUIMessageStream()
     : hasAsyncIterable
-      ? (await import("ai")).createUIMessageStream({
+      ? createAgentUIMessageStream({
           execute: async ({ writer }) => await writeEventsToUiMessageStream(writer, rendered),
         })
-      : (await import("ai")).createUIMessageStream({
+      : createAgentUIMessageStream({
         execute({ writer }) {
           const messageId = crypto.randomUUID()
           writer.write({ type: "start", messageId })
