+++ packages/agent/test/capability-runtime.test.ts
@@ -255,6 +255,7 @@ describe("agent capability runtime", () => {
 
   it("runs model-backed capability lifecycle once per agent run", async () => {
     vi.doMock("ai", () => ({
+      jsonSchema: vi.fn(schema => schema),
       ToolLoopAgent: class {
         async generate() {
           return { text: "ok" }
@@ -285,6 +286,7 @@ describe("agent capability runtime", () => {
 
   it("does not close model-backed capability contexts twice when cleanup fails", async () => {
     vi.doMock("ai", () => ({
+      jsonSchema: vi.fn(schema => schema),
       ToolLoopAgent: class {
         async generate() {
           return { text: "ok" }
@@ -316,6 +318,7 @@ describe("agent capability runtime", () => {
 
   it("does not run invocation-scoped capability phases during agent resolution", async () => {
     vi.doMock("ai", () => ({
+      jsonSchema: vi.fn(schema => schema),
       ToolLoopAgent: class {
         async generate() {
           return { text: "ok" }
