+++ packages/agent/vite.config.ts
@@ -33,6 +33,7 @@ export default defineConfig({
       "src/runtime/empty-registry.ts",
       "src/runtime/workflow.ts",
       "src/server.ts",
+      "src/server/routes.ts",
       "src/test.ts",
       "src/vercel.ts",
       "src/vite.ts",
