+++ packages/kv/src/index.ts
@@ -1,12 +1,14 @@
 export { kv } from "./runtime/storage.ts"
 export type {
   CloudflareKVStoreConfig,
+  DenoKVStoreConfig,
   FsLiteKVStoreConfig,
   KVDriver,
   KVModuleOptions,
   KVStorage,
   KVStoreConfig,
   ResolvedCloudflareKVStoreConfig,
+  ResolvedDenoKVStoreConfig,
   ResolvedFsLiteKVStoreConfig,
   ResolvedKVModuleOptions,
   ResolvedKVStoreConfig,
